solid improvements, bug fixes to dynamic resizing
