+++ SoundSeer/SoundSeerApp.swift
@@ -2,54 +2,12 @@ import SwiftUI
 
 import LaunchAtLogin
 
-
-
 @main
 struct SoundSeerApp: App {
     @StateObject private var spotifyViewModel: SpotifyViewModel = SpotifyViewModel()
-    @State private var isOccluded = false
-    
-
-//    init() {
-//        spotifyViewModel = SpotifyViewModel($isOccluded)
-//    }
+    @State private var window: NSWindow?
 
     var body: some Scene {
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
-        MenuBarExtra {
-
-        } label: {
-            Image(systemName: "forward.end")
-        }
         MenuBarExtra {
             Button("Next Track", systemImage: "forward.end", action: spotifyViewModel.nextTrack)
                 .labelStyle(.titleAndIcon)
@@ -86,8 +44,20 @@ struct SoundSeerApp: App {
 
             Button("Quit", action: spotifyViewModel.quitSoundSeer)
         } label: {
-            MenuBarExtraLabelView(spotifyViewModel: spotifyViewModel)
-            
+            Group {
+                if spotifyViewModel.playerState != .playing || spotifyViewModel.nowPlaying.isEmpty {
+                    Image(systemName: "ear")
+                } else {
+                    Text(spotifyViewModel.nowPlaying)
+                }
+            }
+            .onReceive(NotificationCenter.default.publisher(for: NSWindow.didChangeOcclusionStateNotification)) { notification in
+                guard let window = notification.object as? NSWindow else { return }
+
+                spotifyViewModel.isVisible = window.occlusionState.contains(.visible)
+                print("isVisible: ", spotifyViewModel.isVisible)
+            }
+            .background(WindowAccessor(window: $window))
         }
 
     }
