+++ SoundSeer/WindowAccessor.swift
@@ -1,0 +1,16 @@
+import SwiftUI
+
+// https://stackoverflow.com/a/63439982
+struct WindowAccessor: NSViewRepresentable {
+    @Binding var window: NSWindow?
+
+    func makeNSView(context: Context) -> NSView {
+        let view = NSView()
+        DispatchQueue.main.async {
+            self.window = view.window   // << right after inserted in window
+        }
+        return view
+    }
+
+    func updateNSView(_ nsView: NSView, context: Context) {}
+}
